Move DTO classes to respective folders + ..
..deleted unused usings
