+++ AnimalAdoption/src/AnimalAdoption.Core/DTO/Identity/LoginDTO.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace AnimalAdoption.Core.DTO
+namespace AnimalAdoption.Core.DTO.Identity
 {
 	public class LoginDTO
 	{
