+++ AnimalAdoption/src/AnimalAdoption.Core/DTO/Request/RequestResponse.cs
@@ -1,7 +1,6 @@
 ﻿using AnimalAdoption.Core.Enums;
-using System.ComponentModel.DataAnnotations;
 
-namespace AnimalAdoption.Core.DTO
+namespace AnimalAdoption.Core.DTO.Request
 {
 	public class RequestResponse
 	{
