+++ AnimalAdoption/src/AnimalAdoption.Core/Domain/Entities/Request.cs
@@ -1,6 +1,5 @@
 ﻿using AnimalAdoption.Core.Enums;
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 
 namespace AnimalAdoption.Core.Domain.Entities
 {
