+++ AnimalAdoption/src/AnimalAdoption.UI/Views/_ViewImports.cshtml
@@ -1,4 +1,7 @@
-﻿@using AnimalAdoption.Core.DTO;
+﻿@using AnimalAdoption.Core.DTO.AnimalProfile;
+@using AnimalAdoption.Core.DTO.ContactForm;
+@using AnimalAdoption.Core.DTO.Identity;
+@using AnimalAdoption.Core.DTO.Request;
 @using AnimalAdoption.Core.Enums;
 @using X.PagedList;
 @using X.PagedList.Mvc.Core;
